remove SAN_META, completely refactor label_map.py
