+++ experiments/exp_def.py
@@ -1,6 +1,6 @@
 import yaml
 from data_utils.vocab import Vocabulary
-from data_utils.label_map import TaskType
+from data_utils.task_def import TaskType
 from data_utils.metrics import Metric
 
 class TaskDefs:
@@ -11,11 +11,13 @@ class TaskDefs:
         data_type_map = {}
         task_type_map = {}
         metric_meta_map = {}
+        enable_san_map = {}
         for task, task_def in self._task_def_dic.items():
             n_class_map[task] = task_def["n_class"]
             data_type_map[task] = task_def["data_type"]
             task_type_map[task] = TaskType[task_def["task_type"]]
             metric_meta_map[task] = tuple(Metric[metric_name] for metric_name in task_def["metric_meta"])
+            enable_san_map[task] = task_def["enable_san"]
             if "labels" in task_def:
                 labels = task_def["labels"]
                 label_mapper = Vocabulary(True)
@@ -28,3 +30,4 @@ class TaskDefs:
         self.data_type_map = data_type_map
         self.task_type_map = task_type_map
         self.metric_meta_map = metric_meta_map
+        self.enable_san_map = enable_san_map
